+++ src/main/php/entryfactory/LogEntryFactory.php
@@ -15,7 +15,7 @@ use stubbles\log\Logger;
  *
  * @api
  */
-interface LogEntryFactory
+abstract class LogEntryFactory
 {
     /**
      * creates a log entry container
@@ -24,7 +24,7 @@ interface LogEntryFactory
      * @param   \stubbles\log\Logger  $logger  logger instance to create log entry container for
      * @return  \stubbles\log\LogEntry
      */
-    public function create(string $target, Logger $logger): LogEntry;
+    abstract public function create(string $target, Logger $logger): LogEntry;
 
     /**
      * recreates given log entry
@@ -34,5 +34,8 @@ interface LogEntryFactory
      * @return  \stubbles\log\LogEntry
      * @since   1.1.0
      */
-    public function recreate(LogEntry $logEntry, Logger $logger): LogEntry;
+    public function recreate(LogEntry $logEntry, Logger $logger): LogEntry
+    {
+        return $logEntry;
+    }
 }