+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
   * Raised minimum required PHP version to 7.0
   * introduced scalar type hints and strict type checking
+  * `stubbles\log\entryfactory\LogEntryFactory` is not an interface any more, but an abstract class
+  * deprecated `stubbles\log\entryfactory\AbstractLogEntryFactory`, extend from `stubbles\log\entryfactory\LogEntryFactory` instead, will be removed with 7.0.0 
 
 
 5.0.0 (2016-06-17)